tina-api/requirements.txt
accelerate==1.3.0
datasets==3.2.0
ffmpeg==1.4
pydub==0.25.1
uvicorn==0.34.0
python-multipart
transformers
fastapi
python-dotenv
azure-cognitiveservices-speech==1.31.0
openai>=1.0.0